add unchecked append_nulls and extend_from_array
