+++ vortex-array/src/builders/fixed_size_list.rs
@@ -174,12 +174,8 @@ impl ArrayBuilder for FixedSizeListBuilder {
 
     /// We define the null value of a fixed-size list of size `m` to be a list of `m` placeholder values.
     ///
-    /// We append `n * m` placeholder values to the underlying `elements` array.
-    ///
-    /// The placeholder values depend on the nullability of the element type:
-    /// - If elements are nullable, we use null values
-    /// - If elements are non-nullable, we use zero values
-    fn append_nulls(&mut self, n: usize) {
+    /// We append `n * m` default values to the underlying `elements` array.
+    unsafe fn append_nulls_unchecked(&mut self, n: usize) {
         assert!(
             self.dtype.is_nullable(),
             "tried to append {n} nulls to a non-nullable array builder"
@@ -193,15 +189,7 @@ impl ArrayBuilder for FixedSizeListBuilder {
 
     /// This will increase the capacity if extending with this `array` would go past the original
     /// capacity.
-    fn extend_from_array(&mut self, array: &dyn Array) -> VortexResult<()> {
-        if !self.dtype.eq_with_nullability_superset(array.dtype()) {
-            vortex_bail!(
-                "tried to extend a builder with `DType` {} with an array with `DType {}",
-                self.dtype,
-                array.dtype()
-            );
-        }
-
+    unsafe fn extend_from_array_unchecked(&mut self, array: &dyn Array) -> VortexResult<()> {
         let fsl = array.to_fixed_size_list()?;
         if fsl.is_empty() {
             return Ok(());
