+++ vortex-array/src/builders/list.rs
@@ -164,12 +164,7 @@ impl<O: OffsetPType> ArrayBuilder for ListBuilder<O> {
         self.nulls.append_n_non_nulls(n);
     }
 
-    fn append_nulls(&mut self, n: usize) {
-        assert!(
-            self.dtype.is_nullable(),
-            "tried to append {n} nulls to a non-nullable array builder"
-        );
-
+    unsafe fn append_nulls_unchecked(&mut self, n: usize) {
         let count = self.value_builder.len();
         for _ in 0..n {
             // A list with a null element is can be a list with a zero-span offset and a validity
@@ -181,15 +176,7 @@ impl<O: OffsetPType> ArrayBuilder for ListBuilder<O> {
         self.nulls.append_n_nulls(n);
     }
 
-    fn extend_from_array(&mut self, array: &dyn Array) -> VortexResult<()> {
-        if !self.dtype.eq_with_nullability_superset(array.dtype()) {
-            vortex_bail!(
-                "tried to extend a builder with `DType` {} with an array with `DType {}",
-                self.dtype,
-                array.dtype()
-            );
-        }
-
+    unsafe fn extend_from_array_unchecked(&mut self, array: &dyn Array) -> VortexResult<()> {
         let list = array.to_list()?;
         if list.is_empty() {
             return Ok(());
