+++ vortex-array/src/builders/null.rs
@@ -49,12 +49,11 @@ impl ArrayBuilder for NullBuilder {
         self.length += n;
     }
 
-    fn append_nulls(&mut self, n: usize) {
+    unsafe fn append_nulls_unchecked(&mut self, n: usize) {
         self.length += n;
     }
 
-    fn extend_from_array(&mut self, array: &dyn Array) -> VortexResult<()> {
-        assert_eq!(array.dtype(), &DType::Null);
+    unsafe fn extend_from_array_unchecked(&mut self, array: &dyn Array) -> VortexResult<()> {
         self.append_nulls(array.len());
         Ok(())
     }
