+++ vortex-array/src/builders/varbinview.rs
@@ -9,7 +9,7 @@ use std::sync::Arc;
 
 use vortex_buffer::{Buffer, BufferMut, ByteBuffer, ByteBufferMut};
 use vortex_dtype::DType;
-use vortex_error::{VortexExpect, VortexResult, vortex_bail};
+use vortex_error::{VortexExpect, VortexResult};
 use vortex_mask::Mask;
 use vortex_utils::aliases::hash_map::{Entry, HashMap};
 
@@ -83,7 +83,6 @@ impl VarBinViewBuilder {
     }
 
     /// Appends a value to the builder.
-    #[inline]
     pub fn append_value<S: AsRef<[u8]>>(&mut self, value: S) {
         self.append_value_view(value.as_ref());
         self.nulls.append_non_null();
@@ -94,7 +93,6 @@ impl VarBinViewBuilder {
     /// # Panics
     ///
     /// This method will panic if the input is `None` and the builder is non-nullable.
-    #[inline]
     pub fn append_option<S: AsRef<[u8]>>(&mut self, value: Option<S>) {
         match value {
             Some(value) => self.append_value(value),
@@ -102,7 +100,6 @@ impl VarBinViewBuilder {
         }
     }
 
-    #[inline]
     fn flush_in_progress(&mut self) {
         if self.in_progress.is_empty() {
             return;
@@ -194,43 +191,25 @@ impl ArrayBuilder for VarBinViewBuilder {
         self
     }
 
-    #[inline]
     fn dtype(&self) -> &DType {
         &self.dtype
     }
 
-    #[inline]
     fn len(&self) -> usize {
         self.nulls.len()
     }
 
-    #[inline]
     fn append_zeros(&mut self, n: usize) {
         self.views_builder.push_n(BinaryView::empty_view(), n);
         self.nulls.append_n_non_nulls(n);
     }
 
-    #[inline]
-    fn append_nulls(&mut self, n: usize) {
-        assert!(
-            self.dtype.is_nullable(),
-            "tried to append {n} nulls to a non-nullable array builder"
-        );
-
+    unsafe fn append_nulls_unchecked(&mut self, n: usize) {
         self.views_builder.push_n(BinaryView::empty_view(), n);
         self.nulls.append_n_nulls(n);
     }
 
-    #[inline]
-    fn extend_from_array(&mut self, array: &dyn Array) -> VortexResult<()> {
-        if !self.dtype.eq_with_nullability_superset(array.dtype()) {
-            vortex_bail!(
-                "tried to extend a builder with `DType` {} with an array with `DType {}",
-                self.dtype,
-                array.dtype()
-            );
-        }
-
+    unsafe fn extend_from_array_unchecked(&mut self, array: &dyn Array) -> VortexResult<()> {
         let array = array.to_varbinview()?;
         self.flush_in_progress();
